+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Constrain `extendedFields` initialization to be `Encodable`.
 * Refactor `TileJSON` type to include extended fields and a `Base` object with the base fields. Remove `ExtendedTileJSON`.
 * Rename `tileJSONVersion` to `tileJSON`.
+* Reorganize.
 
 ## 0.3.0
 * Make `extendedFields` default to empty instead of `nil`.